refactor-removed-all-card-related-logic
