--- src/WalletApp.API/Cards/Requests/CreateCardRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace WalletApp.API.Cards.Requests
-{
-    public record CreateCardRequest(
-        string Number,
-        decimal Balance,
-        decimal Available);
-}
