+++ src/WalletApp.API/Constants/Routes.cs
@@ -5,7 +5,6 @@
         public const string Prefix = "api/";
 
         public const string Users = Prefix + "users";
-        public const string Cards = Prefix + "cards";
         public const string MoneyTransactions = Prefix + "moneyTransactions";
     }
 }
