+++ src/WalletApp.API/Controllers/UserContoller.cs
@@ -1,12 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using WalletApp.API.Cards.Mappers;
-using WalletApp.API.Cards.Requests;
 using WalletApp.API.Constants;
 using WalletApp.API.MoneyTransactions.Mappers;
 using WalletApp.API.MoneyTransactions.Requests;
 using WalletApp.API.Users.Mappers;
 using WalletApp.API.Users.Requests;
-using WalletApp.Domain.Cards.Models;
 using WalletApp.Domain.Common;
 using WalletApp.Domain.Transactions.Models;
 using WalletApp.Domain.Users.Data;
@@ -75,38 +72,15 @@ namespace WalletApp.API.Controllers
             var data = request.AsData();
             var newTransaction = MoneyTransaction.Create(data);
             var user = await _unitOfWork.Users.GetById(newTransaction.UserId);
-            var card = await _unitOfWork.Cards.GetById(user.CardId);
 
             if(newTransaction.Type == TransactionType.Payment)
-                card.ChangeBalance(newTransaction.Amount);
+                user.ChangeBalance(newTransaction.Amount);
             if(newTransaction.Type == TransactionType.Credit)
-                card.ChangeBalance(newTransaction.Amount * (-1));
+                user.ChangeBalance(newTransaction.Amount * (-1));
 
             _unitOfWork.MoneyTransactions.Add(newTransaction);
 
             return newTransaction.Id;
-        }
-
-        [HttpGet("/cards")]
-        public async Task<IReadOnlyCollection<Card>> GetAllCards()
-        {
-            return await _unitOfWork.Cards.GetAllCards();
-        }
-
-        [HttpGet("card")]
-        public async Task<Card> GetCardById(int cardId)
-        {
-            return await _unitOfWork.Cards.GetById(cardId);
-        }
-
-        [HttpPost]
-        public async Task<int> CreateCard(CreateCardRequest request)
-        {
-            var data = request.AsData();
-            var newCard = Card.Create(data);
-            _unitOfWork.Cards.Add(newCard);
-
-            return newCard.Id;
         }
     }
 }
