+++ src/WalletApp.API/Users/Mappers/CreateUserMapper.cs
@@ -9,7 +9,6 @@ namespace WalletApp.API.Users.Mappers
             new(request.Name,
                 request.Email,
                 request.Password,
-                request.DueIsPayed,
-                request.CardId);
+                request.DueIsPayed);
     }
 }
