--- src/WalletApp.Domain/Cards/Data/CardCreateData.cs
@@ -1,6 +1,0 @@
-﻿namespace WalletApp.Domain.Cards.Data
-{
-    public record CardCreateData(
-        string Number,
-        decimal Balance);
-}
