+++ src/WalletApp.Domain/Common/IUnitOfWork.cs
@@ -1,5 +1,4 @@
-﻿using WalletApp.Domain.Cards.Common;
-using WalletApp.Domain.Transactions.Common;
+﻿using WalletApp.Domain.Transactions.Common;
 using WalletApp.Domain.Users.Common;
 
 namespace WalletApp.Domain.Common
@@ -8,7 +7,6 @@ namespace WalletApp.Domain.Common
     {
         IUserRepository Users { get; }
         IMoneyTransactionRepository MoneyTransactions { get; }
-        ICardRepository Cards { get; }
 
         Task Complete();
     }
