+++ src/WalletApp.Domain/WalletApp.Domain.csproj
@@ -15,4 +15,10 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Cards\Common\" />
+    <Folder Include="Cards\Data\" />
+    <Folder Include="Cards\Models\" />
+  </ItemGroup>
+
 </Project>
